+++ components/miot_thermogigro/miot_thermogigro.h
@@ -28,6 +28,8 @@ class MiotThermoGigro : public miot::MiotComponent, public sensor::Sensor {
   void process_temperature_(const miot::BLEObject &obj);
   void process_humidity_(const miot::BLEObject &obj);
   void process_temperature_humidity_(const miot::BLEObject &obj);
+  void process_miaomiaoce_temperature_(const miot::BLEObject &obj);
+  void process_miaomiaoce_humidity_(const miot::BLEObject &obj);
 };
 
 }  // namespace miot_thermogigro